v0.05
i just did some cool shit i dont feel like explaining it im tired

i guess there is a cursor now and also I changed the global reference system and also you can destroy and create blocks

+++ 2d voxel/Assets/Scripts/LightMap.cs
@@ -13,7 +13,7 @@ public static class LightMap
     }
 
 
-    public static void Update(Vector2Int start, int[,] lightData)
+    public static void Update(Vector2Int start, byte[,] lightData)
     {
         int width = lightData.GetLength(0);
         int height = lightData.GetLength(1);
